fix : 채팅 패킷 생성 함수 수정
어제 구현한 채팅 기능에서 패킷 생성 함수 수정
+ errorCodes 1줄 추가
