+++ src/utils/error/errorCodes.js
@@ -9,5 +9,6 @@ export const ErrorCodes = {
   INVALID_PACKET: 10007,
   INVALID_SEQUENCE: 10008,
   GAME_NOT_FOUND: 10009,
+  INVALID_CHAT_TYPE: 10010,
   // 여기다 추가
 };
